move export monitoring plan into exported function
