+++ src/components/export/ExportTab/ExportTab.js
@@ -6,10 +6,7 @@ import ReportingPeriodSelector from "../../ReportingPeriodSelector/ReportingPeri
 
 import ExportTablesContainer from "./ExportTablesContainer";
 import { Preloader } from "@us-epa-camd/easey-design-system";
-import { exportQA } from "../../../utils/api/qaCertificationsAPI";
-import { getUnitIdAndStackPipeIds } from "../../QAImportHistoricalDataPreview/QAImportHistoricalDataPreview";
-import { getMonitoringPlanById } from "../../../utils/api/monitoringPlansApi";
-import { getFacilityById } from "../../../utils/api/facilityApi";
+import { exportMonitoringPlanDownload } from "../../../utils/api/monitoringPlansApi";
 
 const ExportTab = ({
   facility,
@@ -91,18 +88,7 @@ const ExportTab = ({
     let exportFileName
     // export monitoring plan
     if (dataTypes.find((e) => e.name === "monitoring-plan").checked) {
-      const mpRes = await getMonitoringPlanById(selectedConfig.id)
-      const facId = mpRes.data["facId"];
-      const mpName = mpRes.data["name"];
-      const date = new Date();
-      const year = date.getUTCFullYear();
-      const month = date.getUTCMonth() + 1;
-      const day = date.getUTCDate();
-      const fullDateString = `${month}-${day}-${year}`;
-      const facRes = await getFacilityById(facId)
-      const facName = facRes.data["facilityName"];
-      exportFileName = `MP Export - ${facName}, ${mpName} (${fullDateString}).json`;
-      download(JSON.stringify(mpRes.data, null, "\t"), exportFileName);
+      exportMonitoringPlanDownload(selectedConfig.id)
       return;
     }
     // export if preview was clicked
